some more frontend work
